change: Move Error types out from circuit_input_builder
Moved all error type definitions like `OOGError` or `ExecError` to the
`error.rs` file and inter-connected them.
